Move: inline 파일에 있는 타입들 분리
